content/base/test/test_bug383430.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=383430
-->
<head>
  <title>Test for Bug 383430</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=383430">Mozilla Bug 383430</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 383430 **/
netscape.security.PrivilegeManager.enablePrivilege("UniversalXPConnect");

var req = new XMLHttpRequest();
req.mozBackgroundRequest = true;
req.open("GET", window.location.href);
req.send(null);

ok(req.channel.loadGroup == null, "loadGroup is null");
</script>
</pre>
</body>
</html>
